requirements.txt
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.0.12
Django==2.2.16
django-filter==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
flake8==5.0.4
idna==3.3
importlib-metadata==4.2.0
iniconfig==1.1.1
mccabe==0.7.0
mypy==0.971
mypy-extensions==0.4.3
numpy==1.21.6
packaging==21.3
pandas==1.3.5
pluggy==0.13.1
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.1.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.3.0
urllib3==1.26.11
zipp==3.8.1